backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
spotipy==2.23.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7

# Testing
pytest==8.1.1
pytest-mock==3.14.0
pytest-cov==4.1.0
responses==0.24.1
